# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Olimpiadas",
        "description": "Documentación de la API del sistema de Olimpiadas",
        "contact": {
            "email": "admin@olimpiadas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/categoryLevel": {
            "get": {
                "tags": [
                    "Niveles y Categorías"
                ],
                "summary": "Obtener todos los niveles y categorías",
                "operationId": "6ad37665d581470cbcce8c2a7b62bcaf",
                "responses": {
                    "200": {
                        "description": "Lista de niveles y categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Niveles/categoría obtenidos exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryLevel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Niveles y Categorías"
                ],
                "summary": "Crear nuevo nivel/categoría",
                "operationId": "933c79a59b59c9bca46c4220eb6ff59b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryLevel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nivel/categoría creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryLevel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/categoryLevel": {
            "get": {
                "tags": [
                    "Niveles y Categorías - Usuario"
                ],
                "summary": "Obtener niveles disponibles",
                "description": "Obtiene la lista de niveles disponibles para usuarios normales",
                "operationId": "e66cc4fafa51de056c63b46db3c393fa",
                "responses": {
                    "200": {
                        "description": "Lista de niveles obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryLevel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/categoryLevel/{category_id}/{area_id}": {
            "get": {
                "tags": [
                    "Niveles y Categorías - Usuario"
                ],
                "summary": "Obtener categoría específica por ID y área",
                "description": "Obtiene un nivel específico según el área seleccionada",
                "operationId": "fed9e1c64ff2a21b2f87a6f9ec10cfcb",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "area_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryLevel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categoryLevel/{id}": {
            "put": {
                "tags": [
                    "Niveles y Categorías"
                ],
                "summary": "Actualizar nivel/categoría",
                "operationId": "b60c2822b8bf1a96c0973aa3260c27a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryLevel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nivel/categoría actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryLevel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Niveles y Categorías"
                ],
                "summary": "Eliminar nivel/categoría",
                "description": "Elimina un nivel o categoría específica",
                "operationId": "4a2eedd99434430e0e6f3f2e6daef36c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del nivel/categoría a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nivel/categoría eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nivel/categoría eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nivel/categoría no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nivel de categoría no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar el nivel/categoría"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/categoryLevel/area/{area_id}": {
            "get": {
                "tags": [
                    "Niveles y Categorías - Usuario"
                ],
                "summary": "Obtener categorías por área",
                "description": "Obtiene todas las categorías asociadas a un área específica",
                "operationId": "c7ecbddb66b26fc2292f74f97a112b28",
                "parameters": [
                    {
                        "name": "area_id",
                        "in": "path",
                        "description": "ID del área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de categorías del área",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryLevel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron categorías para el área",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No existen categorías relacionadas a esta área"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener los niveles/categoría"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Area": {
                "title": "Area",
                "description": "Modelo de Área de Olimpiada",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Matemáticas"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Área de matemáticas y lógica"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
            "CategoryLevel": {
                "title": "CategoryLevel",
                "description": "Modelo de Nivel/Categoría",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Nivel Básico"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Nivel para estudiantes de primaria"
                    },
                    "area_id": {
                        "type": "integer"
                    },
                    "grado": {
                        "type": "string",
                        "enum": [
                            "primaria",
                            "secundaria"
                        ],
                        "example": "primaria"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "area": {
                        "$ref": "#/components/schemas/Area"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Ingresa tu token JWT (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}